backend/requirements.txt
Django
requests
bs4
redis
python-socketio
eventlet
redis
celery
python-dotenv
Pillow
djangorestframework
drf-yasg
flake8
social-auth-app-django
django-extensions
django-filter==2.4.0